+++ example/kernel_v024_example.dart
@@ -93,6 +93,7 @@ void main(List<String> args) async {
   final smartAccountClient = SmartAccountClient(
     account: account,
     bundler: bundler,
+    publicClient: publicClient,
     paymaster: paymaster,
   );